MOD bulk commit of shell scripts Signed-off-by: Zdenek Styblik <stybla@turnovfree.net>
Despite the fact small commits are better, bulk commit of most of shell scripts.
Aim is to deal with Bashisms, convert to Dash and clean up some Perl stuff along
the way. It is likely more commits will follow, although they're going to be
much much smaller than this one.